chore(progress): relocate shared types and mock data
